Deleted all useless iterations and saved models for no reason (#34)
