--- models/cBertGridNew_1k_20250222_095834.pth
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:cdb01dab50c05c8b3f390de81d0f6bf88876c6085906b2225bff023cbfe01128
-size 4040211
