--- models/cBertGridNew_50_20250222_094104.pth
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:d82bf3714f20fbe7bcc488253dd472225fb0579d238d8bb331f4f1887e152bbe
-size 4040211
