--- models/cBertGrid_1k_20250222_145359.pth
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:601300e4eb31274f9f1ff2cff6787a3e3afdebf5a83d54a11cf54f119bb53a09
-size 4040012
