--- models/cBertGrid_50_20250222_143753.pth
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:65ca9571f5d231b91a5cdea458f593b440ba0344a72804ca2a35f0cadaf66183
-size 4040012
