--- models/cBertSimple_manual_20250122_025516.pth
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:1ded3183bb8e3948bf471b5a71fd56b6e09a251c750522cc2ba2ff69eda6cb31
-size 3984958
