+++ testscr/garbage.py
